requirements.txt
Django==1.10.1
django-filter==0.14.0
djangorestframework==3.4.6
MySQL-python==1.2.5
Pillow==3.3.1
South==1.0.2
wsgiref==0.1.2